fixed most of the comments!
